# openapi/generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DataStack API",
    "version": "1.0.0",
    "description": "DataStack platform REST API"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/v1/alerts/preview": {
      "get": {
        "summary": "Preview alert rule",
        "description": "Evaluate whether a rule would trigger with current data",
        "parameters": [
          {
            "in": "query",
            "name": "ruleId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/upgrade": {
      "post": {
        "summary": "Schedule cluster upgrade",
        "description": "Schedule runtime upgrade for a cluster",
        "parameters": [
          {
            "in": "path",
            "name": "clusterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targetRuntimeVersion": {
                    "type": "string"
                  },
                  "scheduleAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/lineage/column/{catalogName}/{schemaName}/{tableName}/{columnName}": {
      "get": {
        "summary": "Get column lineage",
        "parameters": [
          {
            "in": "path",
            "name": "catalogName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "schemaName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tableName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "columnName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/usage-report": {
      "get": {
        "summary": "Get usage report",
        "description": "Get workspace usage summary for a period",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}